requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
click==8.1.3
Flask==2.2.2
Flask-HTTPAuth==4.7.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
Werkzeug==2.2.2
zipp==3.10.0